+++ src/context.ts
@@ -4,7 +4,7 @@ import {
     JwtCredentials,
     PatCredentials,
 } from "./client/authentication/credentials";
-import { AxiosRestClient } from "./client/https/requests";
+import { AxiosRestClient } from "./client/https/https";
 import { BaseJiraClient, JiraClient } from "./client/jira/jira-client";
 import { XrayClientCloud } from "./client/xray/xray-client-cloud";
 import { ServerClient, XrayClientServer } from "./client/xray/xray-client-server";
